+++ lib/omq/cli/push_pull.rb
@@ -8,7 +8,9 @@ module OMQ
 
 
     class PullRunner < BaseRunner
-      def run_loop(task) = run_recv_logic
+      def run_loop(task)
+        config.parallel ? run_parallel_recv(task) : run_recv_logic
+      end
     end
   end
 end